test: move DTS import test new file
